# PERF: speed up .ix by moving deprecation string out of __init__