@gfyoung 2019-01-13
That makes sense, given that it requires fewer looks to find the string.  Surprising how the big change in performance is though...